Split store selection and pending action slices
