+++ src/store/projectStore.ts
@@ -64,8 +64,6 @@ import {
 } from './helpers/geometry'
 import {
   clipperContourToProfile,
-  featuresFormConnectedOverlapGroup,
-  featuresOverlap,
   flattenFeatureToClipperPath,
   unionClipperPaths,
 } from './helpers/clipping'
@@ -75,7 +73,6 @@ import {
   insertDerivedFeatureTreeEntries,
   normalizeDerivedFeatureNameStem,
   previewOffsetFeatures as previewOffsetFeaturesWithFactory,
-  selectedClosedFeaturesFromIds,
   type DerivedFeatureGroup,
 } from './helpers/derivedFeatures'
 import { idNumericSuffix, nextPlacementSession, nextUniqueGeneratedId, syncIdCounter } from './helpers/ids'
@@ -86,10 +83,12 @@ import {
   normalizeFeatureZRange,
   normalizeTool,
 } from './helpers/normalize'
+import { createPendingActionsSlice } from './slices/pendingActionsSlice'
+import { createPendingCompletionSlice } from './slices/pendingCompletionSlice'
+import { createSelectionSlice, emptySelection, sanitizeSelection } from './slices/selectionSlice'
 import type {
   PendingAddTool,
   ProjectStore,
-  SelectionState,
   SketchInsertTarget,
 } from './types'
 
@@ -2053,95 +2052,6 @@ function projectsEqual(a: Project, b: Project): boolean {
   return JSON.stringify(a) === JSON.stringify(b)
 }
 
-function emptySelection(): SelectionState {
-  return {
-    mode: 'feature',
-    selectedFeatureId: null,
-    selectedFeatureIds: [],
-    selectedNode: null,
-    hoveredFeatureId: null,
-    sketchEditTool: null,
-    activeControl: null,
-  }
-}
-
-function sanitizeSelection(project: Project, selection: SelectionState): SelectionState {
-  const selectedNode = selection.selectedNode
-  const selectedFeatureIds = selection.selectedFeatureIds.filter((featureId) =>
-    project.features.some((feature) => feature.id === featureId)
-  )
-  const selectedFeatureId =
-    selection.selectedFeatureId && selectedFeatureIds.includes(selection.selectedFeatureId)
-      ? selection.selectedFeatureId
-      : selectedFeatureIds.at(-1) ?? null
-
-  if (selectedNode?.type === 'feature') {
-    if (selectedFeatureIds.length === 0 || !selectedFeatureId) {
-      return {
-        ...selection,
-        mode: 'feature',
-        selectedFeatureId: null,
-        selectedFeatureIds: [],
-        selectedNode: null,
-        hoveredFeatureId: null,
-        sketchEditTool: null,
-        activeControl: null,
-      }
-    }
-  }
-
-  const hoveredFeatureId =
-    selection.hoveredFeatureId && project.features.some((feature) => feature.id === selection.hoveredFeatureId)
-      ? selection.hoveredFeatureId
-      : null
-
-  const safeSelectedNode =
-    selectedNode?.type === 'folder'
-      ? project.featureFolders.some((folder) => folder.id === selectedNode.folderId)
-        ? selectedNode
-        : null
-      : selectedNode?.type === 'tab'
-        ? project.tabs.some((tab) => tab.id === selectedNode.tabId)
-          ? selectedNode
-          : null
-      : selectedNode?.type === 'tabs_root'
-        ? selectedNode
-      : selectedNode?.type === 'clamp'
-        ? project.clamps.some((clamp) => clamp.id === selectedNode.clampId)
-          ? selectedNode
-          : null
-      : selectedNode?.type === 'clamps_root'
-        ? selectedNode
-      : selectedNode?.type === 'origin'
-        ? selectedNode
-      : selectedNode?.type === 'backdrop'
-        ? project.backdrop
-          ? selectedNode
-          : null
-      : selectedNode?.type === 'features_root'
-        ? selectedNode
-        : selectedNode
-
-  return {
-    ...selection,
-    mode:
-      selectedFeatureIds.length === 1 && selection.selectedNode?.type === 'feature'
-        ? selection.mode
-        : 'feature',
-    selectedFeatureId,
-    selectedFeatureIds,
-    selectedNode:
-      selectedFeatureId
-        ? { type: 'feature', featureId: selectedFeatureId }
-        : selection.selectedNode?.type === 'feature'
-          ? null
-          : safeSelectedNode,
-    hoveredFeatureId,
-    sketchEditTool: selection.mode === 'sketch_edit' ? selection.sketchEditTool : null,
-    activeControl: null,
-  }
-}
-
 // ============================================================
 // Rule: first feature must always be 'add'
 // The part model is built from the first 'add' solid — subsequent
@@ -2161,19 +2071,34 @@ export function isFirstFeatureValid(features: SketchFeature[]): boolean {
 export const useProjectStore = create<ProjectStore>((set, get) => ({
   project: normalizeProject(newProject()),
   pendingAdd: null,
-  pendingMove: null,
-  pendingTransform: null,
-  pendingOffset: null,
-  pendingShapeAction: null,
   backdropImageLoading: false,
-  sketchEditSession: null,
   history: {
     past: [],
     future: [],
     transactionStart: null,
   },
-
-  selection: emptySelection(),
+  ...createSelectionSlice(set, get, {
+    cloneProject,
+    normalizeProject,
+  }),
+  ...createPendingActionsSlice(set, get),
+  ...createPendingCompletionSlice(set, get, {
+    cloneProject,
+    projectsEqual,
+    translateProfile,
+    translateClamp,
+    translateTab,
+    buildCopiedFeatures,
+    buildCopiedClamps,
+    buildCopiedTabs,
+    resizeBackdropFromReference,
+    rotateBackdropFromReference,
+    resizeFeatureFromReference,
+    rotateFeatureFromReference,
+    previewOffsetFeatures,
+    syncFeatureTreeProject,
+    createDerivedFeature,
+  }),
 
   // ── Project ──────────────────────────────────────────────
 
@@ -3897,488 +3822,6 @@ export const useProjectStore = create<ProjectStore>((set, get) => ({
       },
     })),
 
-  // ── Selection ────────────────────────────────────────────
-
-  selectFeature: (id, additive = false) =>
-    set((s) => {
-      const joinMode = s.pendingShapeAction?.kind === 'join'
-      const cutMode = s.pendingShapeAction?.kind === 'cut'
-      const selectedFeature = id ? s.project.features.find((feature) => feature.id === id) ?? null : null
-
-      if (joinMode) {
-        if (selectedFeature && (!selectedFeature.sketch.profile.closed || selectedFeature.locked)) {
-          return {}
-        }
-
-        const proposedIds =
-          !id
-            ? []
-            : additive
-              ? s.selection.selectedFeatureIds.includes(id)
-                ? s.selection.selectedFeatureIds.filter((featureId) => featureId !== id)
-                : [...s.selection.selectedFeatureIds, id]
-              : [id]
-        const proposedFeatures = proposedIds
-          .map((featureId) => s.project.features.find((feature) => feature.id === featureId) ?? null)
-          .filter((feature): feature is SketchFeature => feature !== null)
-        const nextIds = featuresFormConnectedOverlapGroup(proposedFeatures)
-          ? proposedIds
-          : s.selection.selectedFeatureIds
-        const nextPrimaryId = nextIds.at(-1) ?? null
-
-        return {
-          pendingOffset: null,
-          pendingShapeAction: s.pendingShapeAction ? { ...s.pendingShapeAction, entityIds: nextIds } : null,
-          selection: {
-            ...s.selection,
-            selectedFeatureId: nextPrimaryId,
-            selectedFeatureIds: nextIds,
-            selectedNode: nextPrimaryId ? { type: 'feature', featureId: nextPrimaryId } : null,
-            mode: 'feature',
-            activeControl: null,
-          },
-        }
-      }
-
-      if (cutMode) {
-        const pendingShapeAction = s.pendingShapeAction
-        if (!pendingShapeAction || pendingShapeAction.kind !== 'cut') {
-          return {}
-        }
-
-        if (selectedFeature && (!selectedFeature.sketch.profile.closed || selectedFeature.locked)) {
-          return {}
-        }
-
-        if (!id) {
-          return {
-            pendingOffset: null,
-            pendingShapeAction: { ...pendingShapeAction, cutterId: null, targetIds: [] },
-            selection: {
-              ...s.selection,
-              selectedFeatureId: null,
-              selectedFeatureIds: [],
-              selectedNode: null,
-              mode: 'feature',
-              activeControl: null,
-            },
-          }
-        }
-
-        if (!pendingShapeAction.cutterId) {
-          return {
-            pendingOffset: null,
-            pendingShapeAction: { ...pendingShapeAction, cutterId: id, targetIds: [] },
-            selection: {
-              ...s.selection,
-              selectedFeatureId: id,
-              selectedFeatureIds: [id],
-              selectedNode: { type: 'feature', featureId: id },
-              mode: 'feature',
-              activeControl: null,
-            },
-          }
-        }
-
-        if (id === pendingShapeAction.cutterId) {
-          if (additive) {
-            return {}
-          }
-          return {
-            pendingOffset: null,
-            pendingShapeAction: { ...pendingShapeAction, cutterId: id, targetIds: [] },
-            selection: {
-              ...s.selection,
-              selectedFeatureId: id,
-              selectedFeatureIds: [id],
-              selectedNode: { type: 'feature', featureId: id },
-              mode: 'feature',
-              activeControl: null,
-            },
-          }
-        }
-
-        const cutter = s.project.features.find((feature) => feature.id === pendingShapeAction.cutterId) ?? null
-        if (!cutter || !selectedFeature || !featuresOverlap(cutter, selectedFeature)) {
-          return {}
-        }
-
-        const nextTargetIds = additive
-          ? pendingShapeAction.targetIds.includes(id)
-            ? pendingShapeAction.targetIds.filter((featureId) => featureId !== id)
-            : [...pendingShapeAction.targetIds, id]
-          : [id]
-        const nextSelectedIds = [pendingShapeAction.cutterId, ...nextTargetIds]
-        const nextPrimaryId = nextTargetIds.at(-1) ?? pendingShapeAction.cutterId
-
-        return {
-          pendingOffset: null,
-          pendingShapeAction: { ...pendingShapeAction, targetIds: nextTargetIds },
-          selection: {
-            ...s.selection,
-            selectedFeatureId: nextPrimaryId,
-            selectedFeatureIds: nextSelectedIds,
-            selectedNode: nextPrimaryId ? { type: 'feature', featureId: nextPrimaryId } : null,
-            mode: 'feature',
-            activeControl: null,
-          },
-        }
-      }
-
-      return {
-        pendingOffset: null,
-        pendingShapeAction: null,
-        selection: {
-          ...s.selection,
-          ...(id
-            ? additive
-              ? (() => {
-                  const nextIds = s.selection.selectedFeatureIds.includes(id)
-                    ? s.selection.selectedFeatureIds.filter((featureId) => featureId !== id)
-                    : [...s.selection.selectedFeatureIds, id]
-                  const nextPrimaryId =
-                    nextIds.length === 0
-                      ? null
-                      : s.selection.selectedFeatureId === id && s.selection.selectedFeatureIds.includes(id)
-                        ? nextIds.at(-1) ?? null
-                        : id
-                  return {
-                    selectedFeatureId: nextPrimaryId,
-                    selectedFeatureIds: nextIds,
-                    selectedNode: nextPrimaryId ? { type: 'feature', featureId: nextPrimaryId } : null,
-                  }
-                })()
-              : {
-                  selectedFeatureId: id,
-                  selectedFeatureIds: [id],
-                  selectedNode: { type: 'feature', featureId: id },
-                }
-            : {
-                selectedFeatureId: null,
-                selectedFeatureIds: [],
-                selectedNode: null,
-              }),
-          mode: 'feature',
-          activeControl: null,
-        },
-      }
-    }),
-
-  selectFeatures: (ids) =>
-    set((s) => {
-      const joinMode = s.pendingShapeAction?.kind === 'join'
-      const nextIds = ids.filter((id, index) => {
-        const feature = s.project.features.find((entry) => entry.id === id)
-        if (!feature || ids.indexOf(id) !== index) {
-          return false
-        }
-        return joinMode ? feature.sketch.profile.closed && !feature.locked : true
-      })
-      const validJoinIds =
-        joinMode
-          ? (() => {
-              const nextFeatures = nextIds
-                .map((id) => s.project.features.find((feature) => feature.id === id) ?? null)
-                .filter((feature): feature is SketchFeature => feature !== null)
-              return featuresFormConnectedOverlapGroup(nextFeatures)
-                ? nextIds
-                : s.selection.selectedFeatureIds
-            })()
-          : nextIds
-      const nextPrimaryId = validJoinIds.at(-1) ?? null
-
-      return {
-        pendingOffset: null,
-        pendingShapeAction: joinMode && s.pendingShapeAction ? { ...s.pendingShapeAction, entityIds: validJoinIds } : null,
-        selection: {
-          ...s.selection,
-          selectedFeatureId: nextPrimaryId,
-          selectedFeatureIds: validJoinIds,
-          selectedNode: nextPrimaryId ? { type: 'feature', featureId: nextPrimaryId } : null,
-          mode: 'feature',
-          activeControl: null,
-        },
-      }
-    }),
-
-  selectProject: () =>
-    set((s) => ({
-      pendingOffset: null,
-      pendingShapeAction: null,
-      selection: {
-        ...s.selection,
-        selectedFeatureId: null,
-        selectedFeatureIds: [],
-        selectedNode: { type: 'project' },
-        mode: 'feature',
-        activeControl: null,
-      },
-      sketchEditSession: null,
-    })),
-
-  selectGrid: () =>
-    set((s) => ({
-      pendingOffset: null,
-      pendingShapeAction: null,
-      selection: {
-        ...s.selection,
-        selectedFeatureId: null,
-        selectedFeatureIds: [],
-        selectedNode: { type: 'grid' },
-        mode: 'feature',
-      },
-      sketchEditSession: null,
-    })),
-
-  selectStock: () =>
-    set((s) => ({
-      pendingOffset: null,
-      pendingShapeAction: null,
-      selection: {
-        ...s.selection,
-        selectedFeatureId: null,
-        selectedFeatureIds: [],
-        selectedNode: { type: 'stock' },
-        mode: 'feature',
-      },
-      sketchEditSession: null,
-    })),
-
-  selectOrigin: () =>
-    set((s) => ({
-      pendingOffset: null,
-      pendingShapeAction: null,
-      selection: {
-        ...s.selection,
-        selectedFeatureId: null,
-        selectedFeatureIds: [],
-        selectedNode: { type: 'origin' },
-        mode: 'feature',
-        activeControl: null,
-      },
-      sketchEditSession: null,
-    })),
-
-  selectBackdrop: () =>
-    set((s) => ({
-      pendingOffset: null,
-      pendingShapeAction: null,
-      selection: {
-        ...s.selection,
-        selectedFeatureId: null,
-        selectedFeatureIds: [],
-        selectedNode: { type: 'backdrop' },
-        mode: 'feature',
-        activeControl: null,
-      },
-      sketchEditSession: null,
-    })),
-
-  selectFeaturesRoot: () =>
-    set((s) => ({
-      pendingOffset: null,
-      selection: {
-        ...s.selection,
-        selectedFeatureId: null,
-        selectedFeatureIds: [],
-        selectedNode: { type: 'features_root' },
-        mode: 'feature',
-        activeControl: null,
-      },
-      sketchEditSession: null,
-    })),
-
-  selectTabsRoot: () =>
-    set((s) => ({
-      pendingOffset: null,
-      selection: {
-        ...s.selection,
-        selectedFeatureId: null,
-        selectedFeatureIds: [],
-        selectedNode: { type: 'tabs_root' },
-        mode: 'feature',
-        activeControl: null,
-      },
-      sketchEditSession: null,
-    })),
-
-  selectClampsRoot: () =>
-    set((s) => ({
-      pendingOffset: null,
-      selection: {
-        ...s.selection,
-        selectedFeatureId: null,
-        selectedFeatureIds: [],
-        selectedNode: { type: 'clamps_root' },
-        mode: 'feature',
-        activeControl: null,
-      },
-      sketchEditSession: null,
-    })),
-
-  selectFeatureFolder: (id) =>
-    set((s) => ({
-      pendingOffset: null,
-      selection: {
-        ...s.selection,
-        selectedFeatureId: null,
-        selectedFeatureIds: [],
-        selectedNode: { type: 'folder', folderId: id },
-        mode: 'feature',
-        activeControl: null,
-      },
-      sketchEditSession: null,
-    })),
-
-  selectTab: (id) =>
-    set((s) => ({
-      pendingOffset: null,
-      selection: {
-        ...s.selection,
-        selectedFeatureId: null,
-        selectedFeatureIds: [],
-        selectedNode: { type: 'tab', tabId: id },
-        mode: 'feature',
-        activeControl: null,
-      },
-      sketchEditSession: null,
-    })),
-
-  selectClamp: (id) =>
-    set((s) => ({
-      pendingOffset: null,
-      selection: {
-        ...s.selection,
-        selectedFeatureId: null,
-        selectedFeatureIds: [],
-        selectedNode: { type: 'clamp', clampId: id },
-        mode: 'feature',
-        activeControl: null,
-      },
-      sketchEditSession: null,
-    })),
-
-  hoverFeature: (id) =>
-    set((s) => {
-      if (s.selection.hoveredFeatureId === id) {
-        return {}
-      }
-
-      return {
-        selection: { ...s.selection, hoveredFeatureId: id },
-      }
-    }),
-
-  enterSketchEdit: (id) =>
-    set((s) => ({
-      pendingTransform: null,
-      pendingOffset: null,
-      selection: {
-        ...s.selection,
-        selectedFeatureId: id,
-        selectedFeatureIds: [id],
-        selectedNode: { type: 'feature', featureId: id },
-        mode: 'sketch_edit',
-        sketchEditTool: null,
-        activeControl: null,
-      },
-      sketchEditSession: {
-        entityType: 'feature',
-        entityId: id,
-        snapshot: cloneProject(s.project),
-        pastLength: s.history.past.length,
-      },
-    })),
-
-  enterClampEdit: (id) =>
-    set((s) => ({
-      pendingTransform: null,
-      pendingOffset: null,
-      selection: {
-        ...s.selection,
-        selectedFeatureId: null,
-        selectedFeatureIds: [],
-        selectedNode: { type: 'clamp', clampId: id },
-        mode: 'sketch_edit',
-        sketchEditTool: null,
-        activeControl: null,
-      },
-      sketchEditSession: {
-        entityType: 'clamp',
-        entityId: id,
-        snapshot: cloneProject(s.project),
-        pastLength: s.history.past.length,
-      },
-    })),
-
-  enterTabEdit: (id) =>
-    set((s) => ({
-      pendingTransform: null,
-      pendingOffset: null,
-      selection: {
-        ...s.selection,
-        selectedFeatureId: null,
-        selectedFeatureIds: [],
-        selectedNode: { type: 'tab', tabId: id },
-        mode: 'sketch_edit',
-        sketchEditTool: null,
-        activeControl: null,
-      },
-      sketchEditSession: {
-        entityType: 'tab',
-        entityId: id,
-        snapshot: cloneProject(s.project),
-        pastLength: s.history.past.length,
-      },
-    })),
-
-  applySketchEdit: () =>
-    set((s) => ({
-      selection: { ...s.selection, mode: 'feature', sketchEditTool: null, activeControl: null },
-      sketchEditSession: null,
-    })),
-
-  cancelSketchEdit: () =>
-    set((s) => {
-      if (!s.sketchEditSession) {
-        return {
-          selection: { ...s.selection, mode: 'feature', sketchEditTool: null, activeControl: null },
-          sketchEditSession: null,
-        }
-      }
-
-      const restored = normalizeProject(cloneProject(s.sketchEditSession.snapshot))
-      return {
-        project: restored,
-        selection: {
-          ...sanitizeSelection(restored, s.selection),
-          mode: 'feature',
-          sketchEditTool: null,
-          activeControl: null,
-        },
-        sketchEditSession: null,
-        history: {
-          past: s.history.past.slice(0, s.sketchEditSession.pastLength),
-          future: [],
-          transactionStart: null,
-        },
-      }
-    }),
-
-  setSketchEditTool: (tool) =>
-    set((s) => ({
-      selection: {
-        ...s.selection,
-        sketchEditTool: s.selection.mode === 'sketch_edit' ? tool : null,
-        activeControl: null,
-      },
-    })),
-
-  setActiveControl: (control) =>
-    set((s) => ({
-      selection: { ...s.selection, activeControl: control },
-    })),
-
   moveFeatureControl: (featureId, control, point) =>
     set((s) => {
       const nextProject = {
@@ -5441,376 +4884,6 @@ export const useProjectStore = create<ProjectStore>((set, get) => ({
     set({ pendingAdd: null })
   },
 
-  startMoveFeature: (featureId) =>
-    set((s) => {
-      const featureIds = s.selection.selectedFeatureIds.includes(featureId)
-        ? s.selection.selectedFeatureIds
-        : [featureId]
-      const features = featureIds
-        .map((id) => s.project.features.find((entry) => entry.id === id) ?? null)
-        .filter((feature): feature is SketchFeature => feature !== null)
-      if (features.length !== featureIds.length || features.some((feature) => feature.locked)) {
-        return {}
-      }
-
-      return {
-        pendingAdd: null,
-        sketchEditSession: null,
-        pendingMove: { mode: 'move', entityType: 'feature', entityIds: featureIds, fromPoint: null, toPoint: null, session: nextPlacementSession() },
-        pendingTransform: null,
-        pendingOffset: null,
-        pendingShapeAction: null,
-        selection: {
-          ...s.selection,
-          selectedFeatureId: featureId,
-          selectedFeatureIds: featureIds,
-          selectedNode: { type: 'feature', featureId },
-          mode: 'feature',
-          hoveredFeatureId: null,
-          activeControl: null,
-        },
-      }
-    }),
-
-  startCopyFeature: (featureId) =>
-    set((s) => {
-      const featureIds = s.selection.selectedFeatureIds.includes(featureId)
-        ? s.selection.selectedFeatureIds
-        : [featureId]
-      const features = featureIds
-        .map((id) => s.project.features.find((entry) => entry.id === id) ?? null)
-        .filter((feature): feature is SketchFeature => feature !== null)
-      if (features.length !== featureIds.length) {
-        return {}
-      }
-
-      return {
-        pendingAdd: null,
-        sketchEditSession: null,
-        pendingMove: { mode: 'copy', entityType: 'feature', entityIds: featureIds, fromPoint: null, toPoint: null, session: nextPlacementSession() },
-        pendingTransform: null,
-        pendingOffset: null,
-        pendingShapeAction: null,
-        selection: {
-          ...s.selection,
-          selectedFeatureId: featureId,
-          selectedFeatureIds: featureIds,
-          selectedNode: { type: 'feature', featureId },
-          mode: 'feature',
-          hoveredFeatureId: null,
-          activeControl: null,
-        },
-      }
-    }),
-
-  startResizeFeature: (featureId) =>
-    set((s) => {
-      const featureIds = s.selection.selectedFeatureIds.includes(featureId)
-        ? s.selection.selectedFeatureIds
-        : [featureId]
-      const features = featureIds
-        .map((id) => s.project.features.find((entry) => entry.id === id) ?? null)
-        .filter((feature): feature is SketchFeature => feature !== null)
-      if (features.length !== featureIds.length || features.some((feature) => feature.locked)) {
-        return {}
-      }
-
-      return {
-        pendingAdd: null,
-        pendingMove: null,
-        pendingTransform: { mode: 'resize', entityType: 'feature', entityIds: featureIds, referenceStart: null, referenceEnd: null, session: nextPlacementSession() },
-        pendingOffset: null,
-        pendingShapeAction: null,
-        sketchEditSession: null,
-        selection: {
-          ...s.selection,
-          selectedFeatureId: featureId,
-          selectedFeatureIds: featureIds,
-          selectedNode: { type: 'feature', featureId },
-          mode: 'feature',
-          hoveredFeatureId: null,
-          activeControl: null,
-        },
-      }
-    }),
-
-  startRotateFeature: (featureId) =>
-    set((s) => {
-      const featureIds = s.selection.selectedFeatureIds.includes(featureId)
-        ? s.selection.selectedFeatureIds
-        : [featureId]
-      const features = featureIds
-        .map((id) => s.project.features.find((entry) => entry.id === id) ?? null)
-        .filter((feature): feature is SketchFeature => feature !== null)
-      if (features.length !== featureIds.length || features.some((feature) => feature.locked)) {
-        return {}
-      }
-
-      return {
-        pendingAdd: null,
-        pendingMove: null,
-        pendingTransform: { mode: 'rotate', entityType: 'feature', entityIds: featureIds, referenceStart: null, referenceEnd: null, session: nextPlacementSession() },
-        pendingOffset: null,
-        pendingShapeAction: null,
-        sketchEditSession: null,
-        selection: {
-          ...s.selection,
-          selectedFeatureId: featureId,
-          selectedFeatureIds: featureIds,
-          selectedNode: { type: 'feature', featureId },
-          mode: 'feature',
-          hoveredFeatureId: null,
-          activeControl: null,
-        },
-      }
-    }),
-
-  startMoveBackdrop: () =>
-    set((s) => {
-      if (!s.project.backdrop) {
-        return {}
-      }
-
-      return {
-        pendingAdd: null,
-        sketchEditSession: null,
-        pendingMove: { mode: 'move', entityType: 'backdrop', entityIds: ['backdrop'], fromPoint: null, toPoint: null, session: nextPlacementSession() },
-        pendingTransform: null,
-        pendingOffset: null,
-        pendingShapeAction: null,
-        selection: {
-          ...s.selection,
-          selectedFeatureId: null,
-          selectedFeatureIds: [],
-          selectedNode: { type: 'backdrop' },
-          mode: 'feature',
-          hoveredFeatureId: null,
-          activeControl: null,
-        },
-      }
-    }),
-
-  startResizeBackdrop: () =>
-    set((s) => {
-      if (!s.project.backdrop) {
-        return {}
-      }
-
-      return {
-        pendingAdd: null,
-        pendingMove: null,
-        pendingTransform: { mode: 'resize', entityType: 'backdrop', entityIds: [], referenceStart: null, referenceEnd: null, session: nextPlacementSession() },
-        pendingOffset: null,
-        pendingShapeAction: null,
-        sketchEditSession: null,
-        selection: {
-          ...s.selection,
-          selectedFeatureId: null,
-          selectedFeatureIds: [],
-          selectedNode: { type: 'backdrop' },
-          mode: 'feature',
-          hoveredFeatureId: null,
-          activeControl: null,
-        },
-      }
-    }),
-
-  startRotateBackdrop: () =>
-    set((s) => {
-      if (!s.project.backdrop) {
-        return {}
-      }
-
-      return {
-        pendingAdd: null,
-        pendingMove: null,
-        pendingTransform: { mode: 'rotate', entityType: 'backdrop', entityIds: [], referenceStart: null, referenceEnd: null, session: nextPlacementSession() },
-        pendingOffset: null,
-        pendingShapeAction: null,
-        sketchEditSession: null,
-        selection: {
-          ...s.selection,
-          selectedFeatureId: null,
-          selectedFeatureIds: [],
-          selectedNode: { type: 'backdrop' },
-          mode: 'feature',
-          hoveredFeatureId: null,
-          activeControl: null,
-        },
-      }
-    }),
-
-  startJoinSelectedFeatures: () =>
-    set((s) => {
-      const featureIds = selectedClosedFeaturesFromIds(s.project, s.selection.selectedFeatureIds).map((feature) => feature.id)
-
-      return {
-        pendingAdd: null,
-        pendingMove: null,
-        pendingTransform: null,
-        pendingOffset: null,
-        pendingShapeAction: { kind: 'join', entityIds: featureIds, keepOriginals: false, session: nextPlacementSession() },
-        sketchEditSession: null,
-        selection: {
-          ...s.selection,
-          selectedFeatureId: featureIds.at(-1) ?? null,
-          selectedFeatureIds: featureIds,
-          selectedNode: featureIds.at(-1) ? { type: 'feature', featureId: featureIds.at(-1)! } : null,
-          mode: 'feature',
-          hoveredFeatureId: null,
-          activeControl: null,
-        },
-      }
-    }),
-
-  startCutSelectedFeatures: () =>
-    set((s) => {
-      return {
-        pendingAdd: null,
-        pendingMove: null,
-        pendingTransform: null,
-        pendingOffset: null,
-        pendingShapeAction: { kind: 'cut', cutterId: null, targetIds: [], keepOriginals: false, session: nextPlacementSession() },
-        sketchEditSession: null,
-        selection: {
-          ...s.selection,
-          selectedFeatureId: null,
-          selectedFeatureIds: [],
-          selectedNode: null,
-          mode: 'feature',
-          hoveredFeatureId: null,
-          activeControl: null,
-        },
-      }
-    }),
-
-  startOffsetSelectedFeatures: () =>
-    set((s) => {
-      const featureIds = s.selection.selectedFeatureIds
-      const features = selectedClosedFeaturesFromIds(s.project, featureIds)
-      if (features.length === 0 || features.some((feature) => feature.locked || feature.kind === 'text')) {
-        return {}
-      }
-
-      return {
-        pendingAdd: null,
-        pendingMove: null,
-        pendingTransform: null,
-        pendingShapeAction: null,
-        pendingOffset: { entityIds: featureIds, session: nextPlacementSession() },
-        sketchEditSession: null,
-        selection: {
-          ...s.selection,
-          selectedFeatureId: featureIds.at(-1) ?? null,
-          selectedFeatureIds: featureIds,
-          selectedNode: featureIds.at(-1) ? { type: 'feature', featureId: featureIds.at(-1)! } : null,
-          mode: 'feature',
-          hoveredFeatureId: null,
-          activeControl: null,
-        },
-      }
-    }),
-
-  startMoveClamp: (clampId) =>
-    set((s) => {
-      const clamp = s.project.clamps.find((entry) => entry.id === clampId)
-      if (!clamp) {
-        return {}
-      }
-
-      return {
-        pendingAdd: null,
-        sketchEditSession: null,
-        pendingMove: { mode: 'move', entityType: 'clamp', entityIds: [clampId], fromPoint: null, toPoint: null, session: nextPlacementSession() },
-        pendingTransform: null,
-        pendingOffset: null,
-        selection: {
-          ...s.selection,
-          selectedFeatureId: null,
-          selectedFeatureIds: [],
-          selectedNode: { type: 'clamp', clampId },
-          mode: 'feature',
-          hoveredFeatureId: null,
-          activeControl: null,
-        },
-      }
-    }),
-
-  startCopyClamp: (clampId) =>
-    set((s) => {
-      const clamp = s.project.clamps.find((entry) => entry.id === clampId)
-      if (!clamp) {
-        return {}
-      }
-
-      return {
-        pendingAdd: null,
-        sketchEditSession: null,
-        pendingMove: { mode: 'copy', entityType: 'clamp', entityIds: [clampId], fromPoint: null, toPoint: null, session: nextPlacementSession() },
-        pendingTransform: null,
-        pendingOffset: null,
-        selection: {
-          ...s.selection,
-          selectedFeatureId: null,
-          selectedFeatureIds: [],
-          selectedNode: { type: 'clamp', clampId },
-          mode: 'feature',
-          hoveredFeatureId: null,
-          activeControl: null,
-        },
-      }
-    }),
-
-  startMoveTab: (tabId) =>
-    set((s) => {
-      const tab = s.project.tabs.find((entry) => entry.id === tabId)
-      if (!tab) {
-        return {}
-      }
-
-      return {
-        pendingAdd: null,
-        sketchEditSession: null,
-        pendingMove: { mode: 'move', entityType: 'tab', entityIds: [tabId], fromPoint: null, toPoint: null, session: nextPlacementSession() },
-        pendingTransform: null,
-        pendingOffset: null,
-        selection: {
-          ...s.selection,
-          selectedFeatureId: null,
-          selectedFeatureIds: [],
-          selectedNode: { type: 'tab', tabId },
-          mode: 'feature',
-          hoveredFeatureId: null,
-          activeControl: null,
-        },
-      }
-    }),
-
-  startCopyTab: (tabId) =>
-    set((s) => {
-      const tab = s.project.tabs.find((entry) => entry.id === tabId)
-      if (!tab) {
-        return {}
-      }
-
-      return {
-        pendingAdd: null,
-        sketchEditSession: null,
-        pendingMove: { mode: 'copy', entityType: 'tab', entityIds: [tabId], fromPoint: null, toPoint: null, session: nextPlacementSession() },
-        pendingTransform: null,
-        pendingOffset: null,
-        selection: {
-          ...s.selection,
-          selectedFeatureId: null,
-          selectedFeatureIds: [],
-          selectedNode: { type: 'tab', tabId },
-          mode: 'feature',
-          hoveredFeatureId: null,
-          activeControl: null,
-        },
-      }
-    }),
-
   autoPlaceTabsForOperation: (operationId) =>
     set((s) => {
       const operation = s.project.operations.find((entry) => entry.id === operationId) ?? null
@@ -5856,453 +4929,6 @@ export const useProjectStore = create<ProjectStore>((set, get) => ({
         },
       }
     }),
-
-  cancelPendingMove: () => set({ pendingMove: null }),
-
-  cancelPendingTransform: () => set({ pendingTransform: null }),
-
-  cancelPendingShapeAction: () => set({ pendingShapeAction: null }),
-
-  setPendingShapeActionKeepOriginals: (keepOriginals) =>
-    set((s) => ({
-      pendingShapeAction: s.pendingShapeAction ? { ...s.pendingShapeAction, keepOriginals } : null,
-    })),
-
-  cancelPendingOffset: () => set({ pendingOffset: null }),
-
-  setPendingMoveFrom: (point) =>
-    set((s) => ({
-      pendingMove: s.pendingMove ? { ...s.pendingMove, fromPoint: point } : null,
-    })),
-
-  setPendingMoveTo: (point) =>
-    set((s) => ({
-      pendingMove: s.pendingMove ? { ...s.pendingMove, toPoint: point } : null,
-    })),
-
-  setPendingTransformReferenceStart: (point) =>
-    set((s) => ({
-      pendingTransform: s.pendingTransform ? { ...s.pendingTransform, referenceStart: point } : null,
-    })),
-
-  setPendingTransformReferenceEnd: (point) =>
-    set((s) => ({
-      pendingTransform: s.pendingTransform ? { ...s.pendingTransform, referenceEnd: point } : null,
-    })),
-
-  completePendingMove: (toPoint, copyCount = 1) =>
-    set((s) => {
-      if (!s.pendingMove?.fromPoint) {
-        return {}
-      }
-
-      const { entityIds, entityType, fromPoint, mode } = s.pendingMove
-      const dx = toPoint.x - fromPoint.x
-      const dy = toPoint.y - fromPoint.y
-
-      if (Math.abs(dx) < 1e-9 && Math.abs(dy) < 1e-9) {
-        return { pendingMove: null }
-      }
-
-      const normalizedCopyCount = Math.max(1, Math.floor(copyCount))
-      if (entityType === 'backdrop') {
-        if (!s.project.backdrop || mode !== 'move') {
-          return { pendingMove: null }
-        }
-
-        const nextProject = {
-          ...s.project,
-          backdrop: {
-            ...s.project.backdrop,
-            center: {
-              x: s.project.backdrop.center.x + dx,
-              y: s.project.backdrop.center.y + dy,
-            },
-          },
-          meta: { ...s.project.meta, modified: new Date().toISOString() },
-        }
-
-        if (projectsEqual(nextProject, s.project)) {
-          return { pendingMove: null }
-        }
-
-        return {
-          project: nextProject,
-          pendingMove: null,
-          history: {
-            past: [...s.history.past, cloneProject(s.project)].slice(-100),
-            future: [],
-            transactionStart: null,
-          },
-        }
-      }
-
-      if (entityType === 'feature') {
-        const sourceFeatures = entityIds
-          .map((featureId) => s.project.features.find((feature) => feature.id === featureId) ?? null)
-          .filter((feature): feature is SketchFeature => feature !== null)
-        if (sourceFeatures.length !== entityIds.length) {
-          return { pendingMove: null }
-        }
-
-        const createdFeatures =
-          mode === 'copy'
-            ? buildCopiedFeatures(sourceFeatures, s.project.features, dx, dy, normalizedCopyCount)
-            : []
-
-        const nextProject = {
-          ...s.project,
-          features:
-            mode === 'copy'
-              ? [
-                  ...s.project.features,
-                  ...createdFeatures,
-                ]
-              : s.project.features.map((feature) => {
-                  if (!entityIds.includes(feature.id) || feature.locked) {
-                    return feature
-                  }
-
-                  return {
-                    ...feature,
-                    sketch: {
-                      ...feature.sketch,
-                      profile: translateProfile(feature.sketch.profile, dx, dy),
-                    },
-                  }
-                }),
-          featureTree:
-            mode === 'copy'
-              ? [
-                  ...s.project.featureTree,
-                  ...createdFeatures.map((f) => ({ type: 'feature' as const, featureId: f.id })),
-                ]
-              : s.project.featureTree,
-          meta: { ...s.project.meta, modified: new Date().toISOString() },
-        }
-
-        if (projectsEqual(nextProject, s.project)) {
-          return { pendingMove: null }
-        }
-
-        return {
-          project: nextProject,
-          pendingMove: null,
-          selection:
-            mode === 'copy'
-              ? {
-                  ...s.selection,
-                  selectedFeatureId: createdFeatures.at(-1)?.id ?? s.selection.selectedFeatureId,
-                  selectedFeatureIds: createdFeatures.map((feature) => feature.id),
-                  selectedNode: createdFeatures.at(-1)
-                    ? { type: 'feature', featureId: createdFeatures.at(-1)!.id }
-                    : s.selection.selectedNode,
-                }
-              : s.selection,
-          history: {
-            past: [...s.history.past, cloneProject(s.project)].slice(-100),
-            future: [],
-            transactionStart: null,
-          },
-        }
-      }
-
-      if (entityType === 'tab') {
-        const sourceTabs = entityIds
-          .map((tabId) => s.project.tabs.find((tab) => tab.id === tabId) ?? null)
-          .filter((tab): tab is Tab => tab !== null)
-        if (sourceTabs.length !== entityIds.length) {
-          return { pendingMove: null }
-        }
-
-        const createdTabs =
-          mode === 'copy'
-            ? buildCopiedTabs(sourceTabs, s.project.tabs, s.project, dx, dy, normalizedCopyCount)
-            : []
-
-        const nextProject = {
-          ...s.project,
-          tabs:
-            mode === 'copy'
-              ? [...s.project.tabs, ...createdTabs]
-              : s.project.tabs.map((tab) => (
-                  entityIds.includes(tab.id) ? translateTab(tab, dx, dy) : tab
-                )),
-          meta: { ...s.project.meta, modified: new Date().toISOString() },
-        }
-
-        if (projectsEqual(nextProject, s.project)) {
-          return { pendingMove: null }
-        }
-
-        return {
-          project: nextProject,
-          pendingMove: null,
-          selection:
-            mode === 'copy'
-              ? {
-                  ...s.selection,
-                  selectedFeatureId: null,
-                  selectedFeatureIds: [],
-                  selectedNode: createdTabs.at(-1)
-                    ? { type: 'tab', tabId: createdTabs.at(-1)!.id }
-                    : s.selection.selectedNode,
-                }
-              : s.selection,
-          history: {
-            past: [...s.history.past, cloneProject(s.project)].slice(-100),
-            future: [],
-            transactionStart: null,
-          },
-        }
-      }
-
-      const sourceClamps = entityIds
-        .map((clampId) => s.project.clamps.find((clamp) => clamp.id === clampId) ?? null)
-        .filter((clamp): clamp is Clamp => clamp !== null)
-      if (sourceClamps.length !== entityIds.length) {
-        return { pendingMove: null }
-      }
-
-      const createdClamps =
-        mode === 'copy'
-          ? buildCopiedClamps(sourceClamps, s.project.clamps, s.project, dx, dy, normalizedCopyCount)
-          : []
-
-      const nextProject = {
-        ...s.project,
-        clamps:
-          mode === 'copy'
-            ? [...s.project.clamps, ...createdClamps]
-            : s.project.clamps.map((clamp) => (
-                entityIds.includes(clamp.id) ? translateClamp(clamp, dx, dy) : clamp
-              )),
-        meta: { ...s.project.meta, modified: new Date().toISOString() },
-      }
-
-      if (projectsEqual(nextProject, s.project)) {
-        return { pendingMove: null }
-      }
-
-      return {
-        project: nextProject,
-        pendingMove: null,
-        selection:
-          mode === 'copy'
-            ? {
-                ...s.selection,
-                selectedFeatureId: null,
-                selectedFeatureIds: [],
-                selectedNode: createdClamps.at(-1)
-                  ? { type: 'clamp', clampId: createdClamps.at(-1)!.id }
-                  : s.selection.selectedNode,
-              }
-            : s.selection,
-        history: {
-          past: [...s.history.past, cloneProject(s.project)].slice(-100),
-          future: [],
-          transactionStart: null,
-        },
-      }
-    }),
-
-  completePendingTransform: (previewPoint) =>
-    set((s) => {
-      const pendingTransform = s.pendingTransform
-      if (!pendingTransform?.referenceStart || !pendingTransform.referenceEnd) {
-        return {}
-      }
-
-      if (pendingTransform.entityType === 'backdrop') {
-        if (!s.project.backdrop) {
-          return { pendingTransform: null }
-        }
-
-        const nextBackdrop =
-          pendingTransform.mode === 'resize'
-            ? resizeBackdropFromReference(s.project.backdrop, pendingTransform.referenceStart, pendingTransform.referenceEnd, previewPoint)
-            : rotateBackdropFromReference(s.project.backdrop, pendingTransform.referenceStart, pendingTransform.referenceEnd, previewPoint)
-
-        if (!nextBackdrop) {
-          return { pendingTransform: null }
-        }
-
-        const nextProject = {
-          ...s.project,
-          backdrop: nextBackdrop,
-          meta: { ...s.project.meta, modified: new Date().toISOString() },
-        }
-
-        if (projectsEqual(nextProject, s.project)) {
-          return { pendingTransform: null }
-        }
-
-        return {
-          project: nextProject,
-          pendingTransform: null,
-          history: {
-            past: [...s.history.past, cloneProject(s.project)].slice(-100),
-            future: [],
-            transactionStart: null,
-          },
-        }
-      }
-
-      const sourceFeatures = pendingTransform.entityIds
-        .map((featureId) => s.project.features.find((feature) => feature.id === featureId) ?? null)
-        .filter((feature): feature is SketchFeature => feature !== null)
-      if (sourceFeatures.length !== pendingTransform.entityIds.length) {
-        return { pendingTransform: null }
-      }
-
-      const transformedFeatures = new Map<string, SketchFeature>()
-      for (const feature of sourceFeatures) {
-        const transformed =
-          pendingTransform.mode === 'resize'
-            ? resizeFeatureFromReference(feature, pendingTransform.referenceStart, pendingTransform.referenceEnd, previewPoint)
-            : rotateFeatureFromReference(feature, pendingTransform.referenceStart, pendingTransform.referenceEnd, previewPoint)
-        if (!transformed) {
-          return { pendingTransform: null }
-        }
-        transformedFeatures.set(feature.id, transformed)
-      }
-
-      const nextProject = {
-        ...s.project,
-        features: s.project.features.map((feature) => transformedFeatures.get(feature.id) ?? feature),
-        meta: { ...s.project.meta, modified: new Date().toISOString() },
-      }
-
-      if (projectsEqual(nextProject, s.project)) {
-        return { pendingTransform: null }
-      }
-
-      return {
-        project: nextProject,
-        pendingTransform: null,
-        history: {
-          past: [...s.history.past, cloneProject(s.project)].slice(-100),
-          future: [],
-          transactionStart: null,
-        },
-      }
-    }),
-
-  completePendingOffset: (distance) => {
-    const state = get()
-    if (!state.pendingOffset) {
-      return []
-    }
-
-    const createdFeatures = previewOffsetFeatures(state.project, state.pendingOffset.entityIds, distance)
-    if (createdFeatures.length === 0) {
-      set({ pendingOffset: null })
-      return []
-    }
-
-    set((s) => {
-      const nextProject = syncFeatureTreeProject({
-        ...s.project,
-        features: [...s.project.features, ...createdFeatures],
-        meta: { ...s.project.meta, modified: new Date().toISOString() },
-      })
-      const createdIds = createdFeatures.map((feature) => feature.id)
-      const primaryId = createdIds.at(-1) ?? null
-      return {
-        project: nextProject,
-        pendingOffset: null,
-        selection: {
-          ...s.selection,
-          selectedFeatureId: primaryId,
-          selectedFeatureIds: createdIds,
-          selectedNode: primaryId ? { type: 'feature', featureId: primaryId } : null,
-          mode: 'feature',
-          activeControl: null,
-        },
-        history: {
-          past: [...s.history.past, cloneProject(s.project)].slice(-100),
-          future: [],
-          transactionStart: null,
-        },
-      }
-    })
-
-    return createdFeatures.map((feature) => feature.id)
-  },
-
-  completePendingShapeAction: () => {
-    const state = get()
-    const pendingShapeAction = state.pendingShapeAction
-    if (!pendingShapeAction) {
-      return []
-    }
-
-    if (pendingShapeAction.kind === 'join') {
-      if (pendingShapeAction.entityIds.length < 2) {
-        return []
-      }
-
-      state.selectFeatures(pendingShapeAction.entityIds)
-      const result = get().mergeSelectedFeatures(pendingShapeAction.keepOriginals)
-      set({ pendingShapeAction: null })
-      return result
-    }
-
-    if (!pendingShapeAction.cutterId || pendingShapeAction.targetIds.length === 0) {
-      return []
-    }
-
-    const cutter = state.project.features.find((feature) => feature.id === pendingShapeAction.cutterId) ?? null
-    const targets = pendingShapeAction.targetIds
-      .map((featureId) => state.project.features.find((feature) => feature.id === featureId) ?? null)
-      .filter((feature): feature is SketchFeature => feature !== null)
-      .filter((feature) => feature.sketch.profile.closed)
-    if (!cutter || !cutter.sketch.profile.closed || targets.length !== pendingShapeAction.targetIds.length) {
-      return []
-    }
-
-    const createdGroups = cutFeaturesByCutterGrouped(state.project, cutter, targets, createDerivedFeature)
-    const createdFeatures = createdGroups.flatMap((group) => group.features)
-    if (createdFeatures.length === 0) {
-      set({ pendingShapeAction: null })
-      return []
-    }
-
-    set((s) => {
-      const idsToReplace = new Set(
-        pendingShapeAction.keepOriginals
-          ? []
-          : pendingShapeAction.targetIds,
-      )
-      const nextProject = syncFeatureTreeProject({
-        ...s.project,
-        features: insertDerivedFeaturesAfterSources(s.project.features, createdGroups, idsToReplace),
-        featureTree: insertDerivedFeatureTreeEntries(s.project.featureTree, s.project.features, createdGroups, idsToReplace),
-        meta: { ...s.project.meta, modified: new Date().toISOString() },
-      })
-      const createdIds = createdFeatures.map((feature) => feature.id)
-      const primaryId = createdIds.at(-1) ?? null
-      return {
-        project: nextProject,
-        pendingShapeAction: null,
-        selection: {
-          ...s.selection,
-          selectedFeatureId: primaryId,
-          selectedFeatureIds: createdIds,
-          selectedNode: primaryId ? { type: 'feature', featureId: primaryId } : null,
-          mode: 'feature',
-          activeControl: null,
-        },
-        history: {
-          past: [...s.history.past, cloneProject(s.project)].slice(-100),
-          future: [],
-          transactionStart: null,
-        },
-      }
-    })
-
-    return createdFeatures.map((feature) => feature.id)
-  },
 
   // ── Convenience constructors ─────────────────────────────
